Create Updated DTOs using Dev Docs changes | Create TemporaryDTOFieldFinder Package to get example values

+++ src/main/java/com/PhamKornbluhGroup/DTO/ExtendedDTO.java
@@ -1,8 +1,8 @@
 package com.PhamKornbluhGroup.DTO;
 
 public class ExtendedDTO {
-    private String category; // Ex: "weapons"
-    private String[] subcategories; // Ex: "[ "wand" ]"
-    private int prefixes; // Ex: "1" //p Note: Max should be 3 in most cases, but not always due to new heist ring/amulets
-    private int suffixes; // Ex: "3" //p Note: Max should be 3 in most cases, but not always due to new heist ring/amulets
+    private String category;        // Deprecated   // Optional
+    private String[] subcategories; // Deprecated   // Optional
+    private long prefixes;                          // Optional
+    private long suffixes;                          // Optional
 }
